Add temporary input file, type for output file
The adding will delete the input file after the conversion is done. If prompt for input file is configured then the actual file is deleted. I have added so that a temporary file is created and the input file is read into that. This will be the file that is deleted.
When prompt for output file is selected the file type is always pdf. Changed this so that the file type is the type that's configured as OutputExtension.